Update deps and split code
