+++ .gitignore
@@ -1,3 +1,4 @@
 es-log-lag
 bin/
 testing/
+cover.out
